feat: implement previous state in fsm

+++ src/server/TaskManager.cpp
@@ -12,6 +12,8 @@ static void fsm_running_task(Process &process);
 static void fsm_exiting_task(Process &process);
 static void fsm_stopped_task(Process &process);
 
+int counter = 0;
+
 TaskManager::TaskManager(std::unordered_map<std::string, std::vector<Process>>& process_pool, std::mutex& process_pool_mutex)
   : _process_pool(process_pool),
   _process_pool_mutex(process_pool_mutex),
@@ -43,53 +45,61 @@ void TaskManager::work() {
 void TaskManager::fsm(Process &process) {
   const ProgramConfig& config = process.get_program_config();
   Process::status_t status;
+  Process::State next_state;
+  // std::cout << "prev state: " << process.get_previous_state() << std::endl;
+  // std::cout << "curr state: " << process.get_state() << std::endl;
   switch (process.get_state()) {
     case Process::State::Waiting:
-      //std::cout << "[TaskManager] " << config.get_name() << ":" << process.get_pid() << ">(Waiting)" << std::endl;
+      std::cout << "[TaskManager] " << config.get_name() << ":" << process.get_pid() << ">(Waiting)" << std::endl;
       fsm_waiting_task(process, config);
       if (!config.get_autostart()) {
         std::cout << "[TaskManager] " << config.get_name() << ":" << process.get_pid() << " (Waiting)>(Stopped)" << std::endl;
-        process.set_state(Process::State::Stopped);
+        next_state = Process::State::Stopped;
       } else if (config.get_starttime() != 0) {
         std::cout << "[TaskManager] " << config.get_name() << ":" << process.get_pid() << " (Waiting)>(Starting)" << std::endl;
-        process.set_state(Process::State::Starting);
+        next_state = Process::State::Starting;
       } else {
-        std::cout << "[TaskManager] " << config.get_name() << ":" << process.get_pid() << " (Waiting)>(Running)" << std::endl; process.set_state(Process::State::Running);
+        std::cout << "[TaskManager] " << config.get_name() << ":" << process.get_pid() << " (Waiting)>(Running)" << std::endl;
+        next_state = Process::State::Running;
       }
       break;
     case Process::State::Starting:
       fsm_starting_task(process);
+      next_state = Process::State::Starting;
       status = process.get_status();
       if (!status.running) {
-        process.set_state(Process::State::Stopped);
+        next_state = Process::State::Stopped;
       } else if (process.get_runtime() >= config.get_starttime()) {
-        process.set_state(Process::State::Running);
+        next_state = Process::State::Running;
       } else if (process.get_pending_command() == Process::Command::Restart || process.get_pending_command() == Process::Command::Stop) {
-        process.set_state(Process::State::Exiting);
+        next_state = Process::State::Exiting;
       }
       break;
     case Process::State::Running:
       fsm_running_task(process);
+      next_state = Process::State::Running;
       status = process.get_status();
       if (!status.running) {
         std::cout << "[TaskManager] " << config.get_name() << ":" << process.get_pid() << " (Running)>(Stopped)" << std::endl;
-        process.set_state(Process::State::Stopped);
+        next_state = Process::State::Stopped;
       } else if (process.get_pending_command() == Process::Command::Restart || process.get_pending_command() == Process::Command::Stop) {
         std::cout << "[TaskManager] " << config.get_name() << ":" << process.get_pid() << " (Running)>(Exiting)" << std::endl;
-        process.set_state(Process::State::Exiting);
+        next_state = Process::State::Exiting;
       }
       break;
     case Process::State::Exiting:
       fsm_exiting_task(process);
+      next_state = Process::State::Exiting;
       status = process.get_status();
       if (!status.running) {
-        process.set_state(Process::State::Stopped);
+        next_state = Process::State::Stopped;
       } else if (process.get_stoptime() >= config.get_stoptime()) {
-        process.set_state(Process::State::Stopped);
+        next_state = Process::State::Stopped;
       }
       break;
     case Process::State::Stopped:
       fsm_stopped_task(process);
+      next_state = Process::State::Stopped;
       status = process.get_status();
       if (status.running) {
         break;
@@ -98,14 +108,21 @@ void TaskManager::fsm(Process &process) {
           process.check_autorestart()) {
         if (config.get_starttime() == 0) {
           std::cout << "[TaskManager] " << config.get_name() << ":" << process.get_pid() << " (Stopped)>(Running)" << std::endl;
-          process.set_state(Process::State::Running);
+          next_state = Process::State::Running;
         } else {
           std::cout << "[TaskManager] " << config.get_name() << ":" << process.get_pid() << " (Stopped)>(Starting)" << std::endl;
-          process.set_state(Process::State::Starting);
+          next_state = Process::State::Starting;
         }
+      }
+      counter++;
+      std::cout << "counter:" << counter << std::endl;
+      if (counter > 3) {
+        std::exit(0);
       }
       break;
   }
+  process.set_previous_state(process.get_state());
+  process.set_state(next_state);
 }
 
 static void fsm_waiting_task(Process &process, const ProgramConfig &config) {
@@ -115,7 +132,7 @@ static void fsm_waiting_task(Process &process, const ProgramConfig &config) {
 }
 
 static void fsm_starting_task(Process &process) {
-  if (process.get_pid() == -1) {
+  if (process.get_state() != process.get_previous_state()) {
     if (process.get_pending_command() == Process::Command::Start ||
         process.get_pending_command() == Process::Command::Restart) {
       // If a Start/Restart command was issued, reset num_retries and clear the command
@@ -132,7 +149,7 @@ static void fsm_starting_task(Process &process) {
 }
 
 static void fsm_running_task(Process &process) {
-  if (process.get_pid() == -1) {
+  if (process.get_state() != process.get_previous_state()) {
     if (process.get_pending_command() == Process::Command::Start ||
         process.get_pending_command() == Process::Command::Restart) {
       // If a Start/Restart command was issued, clear the command
@@ -148,7 +165,7 @@ static void fsm_exiting_task(Process &process) {
 }
 
 static void fsm_stopped_task(Process &process) {
-  if (process.get_status().running) {
+  if (process.get_previous_state() == Process::State::Exiting && process.get_status().running) {
     process.kill();
     process.update_status();
     return;
